USB hole: Use the Z-axis offset from configuration
Allow overriding the Z-axis offset of the USB hole instead of calculating
it based on the size of the USB hole and the size of the height of the PCB
holder.

Useful when socketing controllers and the offset is not only related to
the thickness of the PCB holder.

Updated the configuration defaults to match the previously calculated value.